requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.10
langchain-community>=0.0.10
google-generativeai>=0.3.2
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pypdf>=4.0.1
tiktoken>=0.5.2 